renamed media-selection-store to multi-media-selection-store
