--- src/Sulu/Bundle/MediaBundle/Resources/js/stores/MediaSelectionStore/index.js
@@ -1,4 +1,0 @@
-// @flow
-import MediaSelectionStore from './MediaSelectionStore';
-
-export default MediaSelectionStore;
