+++ src/Sulu/Bundle/MediaBundle/Resources/js/stores/MultiMediaSelectionStore/index.js
@@ -1,0 +1,4 @@
+// @flow
+import MultiMediaSelectionStore from './MultiMediaSelectionStore';
+
+export default MultiMediaSelectionStore;
