+++ src/Sulu/Bundle/MediaBundle/Resources/js/stores/MultiMediaSelectionStore/types.js
